requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
motor
openai
PyYAML
pydantic-settings
